Automatic tune of Roll
v0 - trying to do an automatic tune of Ki of roll during the run itself based on the difference between pitch error and roll error. this is because we pitch error which is the most tuned and is basically zero and the roll error which is reacting similarly to pitch error but with steady state error, so Ki is needed to remove that.